requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
backports.entry-points-selectable==1.1.1
CacheControl==0.12.10
cachy==0.3.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
cleo==0.8.1
clikit==0.6.2
crashtest==0.3.1
cryptography==36.0.1
discord.py==1.7.3
distlib==0.3.4
filelock==3.4.0
html5lib==1.1
idna==3.3
jeepney==0.7.1
keyring==21.8.0
lockfile==0.12.2
loguru==0.5.3
msgpack==1.0.3
multidict==5.2.0
packaging==20.9
pastel==0.2.1
pexpect==4.8.0
pkginfo==1.8.2
platformdirs==2.4.0
poetry==1.1.12
poetry-core==1.0.7
ptyprocess==0.7.0
pycparser==2.21
pylev==1.4.0
pyparsing==3.0.6
python-dotenv==0.19.2
requests==2.26.0
requests-toolbelt==0.9.1
SecretStorage==3.3.1
shellingham==1.4.0
six==1.16.0
tomlkit==0.7.2
typing_extensions==4.0.1
urllib3==1.26.7
virtualenv==20.10.0
webencodings==0.5.1
yarl==1.7.2
